fix: converted showToolTip into a hook
