+++ apps/meteor/client/views/topup/TopUpView.tsx
@@ -9,7 +9,7 @@ import React, { ReactElement, useEffect, useMemo, useState } from 'react';
 
 import Page from '../../components/Page';
 import ProfileHeader from '../../components/ProfileHeader/ProfileHeader';
-import { useCapitalizeAndJoin } from '../../hooks/useCapitalization';
+import { useCapitalizeAndJoin } from '../../hooks/gso/useCapitalization';
 import Components from './components/Components';
 import './topup.css';
 
@@ -93,8 +93,7 @@ const TopUpView = (): ReactElement => {
 	return (
 		<Page id='topup-page'>
 			{/* @ts-ignore */}
-			<ProfileHeader title={t('gso_topupView_profileHeader')}
-										 handleRouteBack={handleRouteBack} />
+			<ProfileHeader title={t('gso_topupView_profileHeader')} handleRouteBack={handleRouteBack} />
 			<Page.ScrollableContentWithShadow>
 				{/* @ts-ignore */}
 				<h3 style={{ fontSize: '19px', marginBottom: '10px' }}>{t('gso_topupView_title', { currency: inputCurrency })}</h3>
@@ -103,26 +102,26 @@ const TopUpView = (): ReactElement => {
 				<Accordion style={{ margin: '15px 0' }}>
 					{fetchedGateways.length
 						? fetchedGateways.map((gateway) => {
-							if (gateway.cmpClass === undefined || gateway.cmpClass === '') {
-								return <Accordion.Item title={capitalize(gateway._id)} disabled={true} />;
-							}
+								if (gateway.cmpClass === undefined || gateway.cmpClass === '') {
+									return <Accordion.Item title={capitalize(gateway._id)} disabled={true} />;
+								}
 
-							if (gateway.show === false) {
-								return (
-									// @ts-ignore
-									<Accordion.Item title={capitalize(gateway._id)} id={gateway._id} onToggle={onAccordionToggle}>
-										{/* eslint-disable-next-line new-cap */}
-										{Components({
-											id: gateway._id,
-											cmpClass: gateway.cmpClass,
-											capitalize,
-											onAccordionToggle,
-										})}
-									</Accordion.Item>
-								);
-							}
-							return null;
-						})
+								if (gateway.show === false) {
+									return (
+										// @ts-ignore
+										<Accordion.Item title={capitalize(gateway._id)} id={gateway._id} onToggle={onAccordionToggle}>
+											{/* eslint-disable-next-line new-cap */}
+											{Components({
+												id: gateway._id,
+												cmpClass: gateway.cmpClass,
+												capitalize,
+												onAccordionToggle,
+											})}
+										</Accordion.Item>
+									);
+								}
+								return null;
+						  })
 						: 'Loading...'}
 				</Accordion>
 			</Page.ScrollableContentWithShadow>
